Heroku-Binance_MACD/requirements.txt
pushbullet.py==0.12.0
pause==0.3
python_binance==1.0.10
tabulate==0.8.9
pandas_ta==0.2.45b0
pandas==1.2.4
binance==0.3
